requirements.txt
windows-curses
i18nice[YAML]